Better `Show` instances. ...
Legion will sometimes log requests and partition state at debug level,
and these log messages are based off of the `Show` instances of our data
types.  Therefore, we need to make them reasonably efficient, which
means not showing all of the actual bytes of the state data, which
could be huge.